requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
websockets
numpy
pandas
matplotlib
reportlab
gymnasium
openenv-core==0.2.3
trl
transformers
datasets
torch
unsloth
accelerate
openai
requests
google-generativeai
anthropic
aiohttp
redis
python-dotenv
peft
bitsandbytes
